clean code updated.
