+++ lib/global/global-variables.dart
@@ -7,20 +7,21 @@ BehaviorSubject<bool> isLoading$ = BehaviorSubject.seeded(false);
 BehaviorSubject<List<MemberModel>?> member$ = BehaviorSubject.seeded(null);
 
 final config = FitConfig.fromJson(fitConfig);
-final url = Uri.parse('https://4001.hoteladvisor.net');
-String? selectedlang;
+final apiUrl = Uri.parse('https://4001.hoteladvisor.net');
 
+String? selectedlang;
 int? hotelId;
 
-EdgeInsets marginAll5 = EdgeInsets.all(5);
-EdgeInsets marginAll8 = EdgeInsets.all(8);
-EdgeInsets marginAll10 = EdgeInsets.all(10);
-EdgeInsets paddingAll5 = EdgeInsets.all(5);
-EdgeInsets paddingAll8 = EdgeInsets.all(8);
-EdgeInsets paddingAll10 = EdgeInsets.all(10);
-EdgeInsets paddingAll15 = EdgeInsets.all(15);
+EdgeInsets marginAll5 = const EdgeInsets.all(5);
+EdgeInsets marginAll8 = const EdgeInsets.all(8);
+EdgeInsets marginAll10 = const EdgeInsets.all(10);
+EdgeInsets paddingAll5 = const EdgeInsets.all(5);
+EdgeInsets paddingAll8 = const EdgeInsets.all(8);
+EdgeInsets paddingAll10 = const EdgeInsets.all(10);
+EdgeInsets paddingAll15 = const EdgeInsets.all(15);
 
-Decoration borderAndBorderRadius = BoxDecoration(borderRadius: BorderRadius.all(Radius.circular(10)), border: Border.all(color: isDarkMode$.value ? Colors.white : Colors.black87, width: 1));
+Decoration borderAndBorderRadius = BoxDecoration(
+    borderRadius: BorderRadius.all(Radius.circular(10)), border: Border.all(color: isDarkMode$.value ? Colors.white : Colors.black87, width: 1));
 BorderRadius borderRadius8 = BorderRadius.all(Radius.circular(8));
 BorderRadius borderRadius10 = BorderRadius.all(Radius.circular(10));
 BorderRadius borderRadius15 = BorderRadius.all(Radius.circular(15));
