+++ lib/module/my-operation/reservation-create.dart
@@ -61,87 +61,80 @@ class _ReservationCreateState extends State<ReservationCreate> {
     final double W = MediaQuery.of(context).size.width;
 
     return Scaffold(
-      appBar: AppBar(title: Text("Reservation Summary".tr())),
-      body: GestureDetector(
-        onTap: () {
-          FocusScope.of(context).unfocus();
-        },
-        child: SingleChildScrollView(
-          child: Form(
-              key: _formKey,
-              child: Container(
-                padding: paddingAll10,
-                height: H * 0.89,
-                child: Column(
-                  children: [
-                    Column(mainAxisAlignment: MainAxisAlignment.start, crossAxisAlignment: CrossAxisAlignment.start, children: [
-                      Text("Service Information".tr(), style: kMontserrat17),
-                      const Divider(color: Colors.black45),
-                      Row(
-                          mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                          children: [Text("Service Name".tr(), style: kProxima17), Text(widget.spaService.product, style: kProxima17)]),
-                      Row(mainAxisAlignment: MainAxisAlignment.spaceBetween, children: [
-                        Text("Service Price".tr(), style: kProxima17),
-                        Text("${widget.spaService.price} ${widget.spaService.currency}", style: kProxima17)
-                      ]),
-                      Row(mainAxisAlignment: MainAxisAlignment.spaceBetween, children: [
-                        Text("Reservation Date".tr(), style: kProxima17),
-                        Text(DateFormat("dd MMM yyyy").format(widget.resStart), style: kProxima17)
-                      ]),
-                      Row(
-                          mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                          children: [Text("Reservation Time".tr(), style: kProxima17), Text(widget.selectedHours.toString(), style: kProxima17)]),
-                      SizedBox(height: W / 30),
-                      Text("Person Information".tr(), style: kMontserrat17),
-                      const Divider(color: Colors.black45),
-                      CTextFormField(
-                        _nameAndSurname,
-                        "Name Surname".tr(),
-                        keyboardType: TextInputType.text,
-                        textInputAction: TextInputAction.go,
-                        onchange: (v) {
-                          member$.value?.first.profile.fullname = v;
-                        },
-                        validator: (value) {
-                          if (value.isEmpty) {
-                            return 'Please enter your name and surname.'.tr();
-                          }
-                          return null;
-                        },
-                      ),
-                      SizedBox(height: W / 40),
-                      CTextFormField(_email, "Email".tr(), keyboardType: TextInputType.emailAddress, textInputAction: TextInputAction.go,
-                          onchange: (v) {
-                        member$.value?.first.profile.email = v;
-                      }, validator: (value) {
-                        if (value.isEmpty) {
-                          return 'Please enter your email address.'.tr();
-                        }
-                        return null;
-                      }),
-                      SizedBox(height: W / 40),
-                      CTextFormField(_phone, "Phone Number".tr(), keyboardType: TextInputType.phone, textInputAction: TextInputAction.done,
-                          validator: (value) {
-                        if (value.isEmpty) {
-                          return 'Please enter your phone number.'.tr();
-                        }
-                        return null;
-                      }, onchange: (v) {
-                        member$.value?.first.profile.phone = v;
-                      }),
-                      SizedBox(height: W / 40)
-                    ]),
-                    const Spacer(),
-                    Padding(
-                      padding: EdgeInsets.only(bottom: MediaQuery.of(context).padding.bottom),
-                      child: CButton(title: "Complete The Reservation".tr(), func: reservationCreate, width: W),
-                    )
-                  ],
-                ),
-              )),
-        ),
-      ),
-    );
+        appBar: AppBar(title: Text("Reservation Summary".tr())),
+        body: GestureDetector(
+            onTap: () => FocusScope.of(context).unfocus(),
+            child: SingleChildScrollView(
+              child: Form(
+                  key: _formKey,
+                  child: Container(
+                      padding: paddingAll10,
+                      height: H * 0.89,
+                      child: Column(children: [
+                        Column(mainAxisAlignment: MainAxisAlignment.start, crossAxisAlignment: CrossAxisAlignment.start, children: [
+                          Text("Service Information".tr(), style: kMontserrat17),
+                          const Divider(color: Colors.black45),
+                          Row(
+                              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                              children: [Text("Service Name".tr(), style: kProxima17), Text(widget.spaService.product, style: kProxima17)]),
+                          Row(mainAxisAlignment: MainAxisAlignment.spaceBetween, children: [
+                            Text("Service Price".tr(), style: kProxima17),
+                            Text("${widget.spaService.price} ${widget.spaService.currency}", style: kProxima17)
+                          ]),
+                          Row(mainAxisAlignment: MainAxisAlignment.spaceBetween, children: [
+                            Text("Reservation Date".tr(), style: kProxima17),
+                            Text(DateFormat("dd MMM yyyy").format(widget.resStart), style: kProxima17)
+                          ]),
+                          Row(
+                              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                              children: [Text("Reservation Time".tr(), style: kProxima17), Text(widget.selectedHours.toString(), style: kProxima17)]),
+                          SizedBox(height: W / 30),
+                          Text("Person Information".tr(), style: kMontserrat17),
+                          const Divider(color: Colors.black45),
+                          CTextFormField(
+                            _nameAndSurname,
+                            "Name Surname".tr(),
+                            keyboardType: TextInputType.text,
+                            textInputAction: TextInputAction.go,
+                            onchange: (v) {
+                              member$.value?.first.profile.fullname = v;
+                            },
+                            validator: (value) {
+                              if (value.isEmpty) {
+                                return 'Please enter your name and surname.'.tr();
+                              }
+                              return null;
+                            },
+                          ),
+                          SizedBox(height: W / 40),
+                          CTextFormField(_email, "Email".tr(), keyboardType: TextInputType.emailAddress, textInputAction: TextInputAction.go,
+                              onchange: (v) {
+                            member$.value?.first.profile.email = v;
+                          }, validator: (value) {
+                            if (value.isEmpty) {
+                              return 'Please enter your email address.'.tr();
+                            }
+                            return null;
+                          }),
+                          SizedBox(height: W / 40),
+                          CTextFormField(_phone, "Phone Number".tr(), keyboardType: TextInputType.phone, textInputAction: TextInputAction.done,
+                              validator: (value) {
+                            if (value.isEmpty) {
+                              return 'Please enter your phone number.'.tr();
+                            }
+                            return null;
+                          }, onchange: (v) {
+                            member$.value?.first.profile.phone = v;
+                          }),
+                          SizedBox(height: W / 40)
+                        ]),
+                        const Spacer(),
+                        Padding(
+                          padding: EdgeInsets.only(bottom: MediaQuery.of(context).padding.bottom),
+                          child: CButton(title: "Complete The Reservation".tr(), func: reservationCreate, width: W),
+                        )
+                      ]))),
+            )));
   }
 
   Future<dynamic> reservationSuccesfullWidget(
@@ -172,22 +165,16 @@ class _ReservationCreateState extends State<ReservationCreate> {
                             StreamBuilder(
                                 stream: service.resId$.stream,
                                 builder: (context, snapshot) {
-                                  return Row(
-                                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                                    children: [
-                                      Text("Reservation Number".tr(), style: kProxima17),
-                                      Text(service.resId$.value.toString(), style: kProxima17)
-                                    ],
-                                  );
+                                  return Row(mainAxisAlignment: MainAxisAlignment.spaceBetween, children: [
+                                    Text("Reservation Number".tr(), style: kProxima17),
+                                    Text(service.resId$.value.toString(), style: kProxima17)
+                                  ]);
                                 }),
                             SizedBox(width: W / 60),
-                            Row(
-                              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                              children: [
-                                Text("Reservation Date".tr(), style: kProxima17),
-                                Text(DateFormat("dd MMM yyyy").format(resStart), style: kProxima17)
-                              ],
-                            ),
+                            Row(mainAxisAlignment: MainAxisAlignment.spaceBetween, children: [
+                              Text("Reservation Date".tr(), style: kProxima17),
+                              Text(DateFormat("dd MMM yyyy").format(resStart), style: kProxima17)
+                            ]),
                             SizedBox(width: W / 60),
                             Row(
                               mainAxisAlignment: MainAxisAlignment.spaceBetween,
@@ -237,7 +224,7 @@ class _ReservationCreateState extends State<ReservationCreate> {
                                     Navigator.pushReplacement(
                                         context,
                                         PageRouteBuilder(
-                                          pageBuilder: (context, animation, secondaryAnimation) => MyOperations(),
+                                          pageBuilder: (context, animation, secondaryAnimation) => const MyOperations(),
                                         ));
                                   },
                                   child: Container(
