+++ lib/widget/CButton.dart
@@ -22,8 +22,6 @@ class CButton extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    double H = MediaQuery.of(context).size.height;
-    double W = MediaQuery.of(context).size.width;
     return StreamBuilder(
         stream: isDarkMode$.stream,
         builder: (context, snapshot) {
@@ -33,10 +31,15 @@ class CButton extends StatelessWidget {
                 width: width ?? MediaQuery.of(context).size.width * 0.43,
                 height: height ?? MediaQuery.of(context).size.height * 0.06,
                 decoration: BoxDecoration(
-                    borderRadius: const BorderRadius.all(Radius.circular(10)), color: isBorder ? Colors.transparent : config.primaryColor, border: Border.all(color: config.primaryColor, width: 1)),
+                    borderRadius: const BorderRadius.all(Radius.circular(10)),
+                    color: isBorder ? Colors.transparent : config.primaryColor,
+                    border: Border.all(color: config.primaryColor, width: 1)),
                 child: Container(
                   alignment: Alignment.center,
-                  child: Text(title.tr(), style: const TextStyle(color: Colors.white, fontFamily: "Montserrat", fontSize: 18), textAlign: TextAlign.center),
+                  child: isLoading$.value
+                      ? CircularProgressIndicator(color: Colors.white)
+                      : Text(title.tr(),
+                          style: const TextStyle(color: Colors.white, fontFamily: "Montserrat", fontSize: 18), textAlign: TextAlign.center),
                 ),
               ));
         });
